AI Assistant: Move useTextContentFromSelectedBlocks to hooks folder
