+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/lib/utils/block-content.ts
@@ -3,14 +3,11 @@
  */
 import { getBlockContent } from '@wordpress/blocks';
 import { serialize } from '@wordpress/blocks';
-import { select, useSelect } from '@wordpress/data';
+import { select } from '@wordpress/data';
 import TurndownService from 'turndown';
-import type { BlockEditorStore } from '../../types';
 
 // Turndown instance
 const turndownService = new TurndownService();
-
-const HTML_JOIN_CHARACTERS = '<br />';
 
 /**
  * Returns partial content from the beginning of the post
@@ -50,61 +47,6 @@ export function getContentFromBlocks(): string {
 	}
 
 	return turndownService.turndown( serialize( blocks ) );
-}
-
-type GetTextContentFromBlocksProps = {
-	count: number;
-	clientIds: string[];
-	content: string;
-};
-
-/**
- * Returns the text content from all selected blocks.
- *
- * @returns {GetTextContentFromBlocksProps} The text content.
- */
-
-export function useTextContentFromSelectedBlocks(): GetTextContentFromBlocksProps {
-	const clientIds = useSelect(
-		selectFromHook =>
-			(
-				selectFromHook( 'core/block-editor' ) as BlockEditorStore[ 'selectors' ]
-			 ).getSelectedBlockClientIds(),
-		[]
-	);
-
-	const blocks = useSelect(
-		selectFromHook =>
-			(
-				selectFromHook( 'core/block-editor' ) as BlockEditorStore[ 'selectors' ]
-			 ).getBlocksByClientId( clientIds ),
-		[ clientIds ]
-	);
-
-	const defaultContent = {
-		count: 0,
-		clientIds: [],
-		content: '',
-	};
-
-	if ( ! clientIds?.length ) {
-		return defaultContent;
-	}
-
-	if ( ! blocks?.length ) {
-		return defaultContent;
-	}
-
-	return {
-		count: blocks.length,
-		clientIds,
-		content: blocks
-			? blocks
-					.filter( block => block !== null && block !== undefined ) // Safeguard against null or undefined blocks
-					.map( block => getBlockTextContent( block.clientId ) )
-					.join( HTML_JOIN_CHARACTERS )
-			: '',
-	};
 }
 
 /**
